SSO: Load JS asynchronously.
This commit loads the JS that sets the auth cookies asynchronously.

Mercator's SSO works by loading this script only for anonymous users
and sends a request to the main site to see if the user is actually
logged in. If logged in, the JS will redirect back to the mapped
domain, set the logged-in cookies for the mapped domain and redirect
the user back to the page that was being viewed.

However, for visitors that are not part of the site, this JS file adds
unnecessary load time and needs to complete running before other
elements of the page can load. By running this JS asynchronously, other
parts of the page can load at the same time and will speed up the
pageload time. (Load time can be sped up by a few seconds depending
on the server.)

This commit also removes the check to see if the test cookie is
written before the SSO function runs on the mapped domain. If a user
is logged-in from the main site that means cookies are already set, so
no need to do the test cookie check, run the SSO function immediately.